requirements/base.txt
asgiref==3.8.1
attrs==25.1.0
Django==5.1.6
django-cleanup==9.0.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
exceptiongroup==1.2.2
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
model-bakery==1.20.4
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pytest==8.3.5
pytest-django==4.10.0
python-decouple==3.8
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.22.3
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
pika==1.3.2
redis==5.0.8
celery==5.3.4
django-redis==5.4.0
